Fix windows CI issue

+++ lib/validator/formchecker.cpp
@@ -351,10 +351,10 @@ Expect<void> FormChecker::checkInstr(const AST::Instruction &Instr) {
         auto [TryCnt, CatchCnt] = countCtrlStackType(*D + 1);
         auto &NonConstInstr = const_cast<AST::Instruction &>(Instr);
         if (*D + 1 < CtrlStack.size()) {
-          NonConstInstr.setTryBlockVSize(Locals.size() +
+          NonConstInstr.setTryBlockVSize(static_cast<uint32_t>(Locals.size()) +
                                          CtrlStack[*D + 1].Height);
         } else {
-          NonConstInstr.setTryBlockVSize(Locals.size() + ValStack.size());
+          NonConstInstr.setTryBlockVSize(static_cast<uint32_t>(Locals.size()) + ValStack.size());
         }
         // The try block itself may push an additional handler to stack
         NonConstInstr.setTryBlockHOffset(TryCnt);
@@ -373,7 +373,7 @@ Expect<void> FormChecker::checkInstr(const AST::Instruction &Instr) {
       return Unexpect(Res);
     }
     if (Instr.getOpCode() == OpCode::Try) {
-      const_cast<AST::Instruction &>(Instr).setTryBlockParamNum(T1.size());
+      const_cast<AST::Instruction &>(Instr).setTryBlockParamNum(static_cast<uint32_t>(T1.size()));
     }
     // Pop and check [t1*]
     if (auto Res = popTypes(T1); !Res) {
